+++ flapp/forms.py
@@ -47,3 +47,25 @@ class TagForm(forms.ModelForm):
 	def __init__(self, *args, **kwargs):
 		super().__init__(*args, **kwargs)
 		self.fields['name'].widget.attrs.update({'class': 'form-control'})
+
+
+def email_for_subscribers(self):
+	subscribers = [s.email for s in Subscriber.objects.all()]
+	from_email = settings.EMAIL_HOST_USER
+	to_email = subscribers
+	msg = f"{self.title}\nNew Post is uploaded."
+	sub = "New Article"
+	print("added")
+	message = Mail(
+		from_email=from_email,
+		to_emails=to_email,
+		subject=sub,
+		html_content=msg)
+	try:
+		sg = SendGridAPIClient(settings.SENDGRID_API_KEY)
+		response = sg.send(message)
+		print(response.status_code)
+		print(response.body)
+		print(response.headers)
+	except Exception as e:
+		print(e)
